TICK-0006: fixed bags
