+++ app/logical_blocks/created_match.py
@@ -5,7 +5,7 @@ from aiogram.fsm.context import FSMContext
 # Импортируйте модули, которые используются внутри функций
 import ClassesStatesMachine.SG as SG
 import app.DatabaseWork.master as master_db
-import app.keyboards as kb
+import app.keyboards.created_match as kb
 
 from app.message_designer.deletezer import delete_message
